refactor(frontend): split the Pulls screen from the pull request it shows
Pulls.tsx held six components in 732 lines: the screen, the pull request in
full, its merge dialog, the skeleton, the empty state and a notice. The screen
and the view have genuinely separate jobs — which pull request is in view and
what an action there means for the checkout it lives in, against the header
and the four tabs that read one — and the file gave no seam between them.

Each component moves to a file named after it. The merge dialog takes its seed
messages with it as mergeEditFor, which is where the "what GitHub would have
written" decision belongs; the view keeps the in-flight state of the actions it
owns and nothing else.

Pulls.tsx: 732 -> 255 lines, and no file in the folder is now above 330.
